+++ federation/entities/activitypub/mappers.py
@@ -1,5 +1,5 @@
 import logging
-from typing import List, Callable, Dict, Union
+from typing import List, Callable, Dict, Union, Optional
 
 from federation.entities.activitypub.constants import NAMESPACE_PUBLIC
 from federation.entities.activitypub.entities import (
@@ -7,7 +7,7 @@ from federation.entities.activitypub.entities import (
     ActivitypubRetraction)
 from federation.entities.base import Follow, Profile, Accept, Post, Comment, Retraction
 from federation.entities.mixins import BaseEntity
-from federation.types import UserType
+from federation.types import UserType, ReceiverVariant
 
 logger = logging.getLogger("federation")
 
@@ -26,7 +26,7 @@ MAPPINGS = {
 
 def element_to_objects(payload: Dict) -> List:
     """
-    Transform an Element to a list of entities recursively.
+    Transform an Element to a list of entities.
     """
     entities = []
     if payload.get('type') == "Delete":
@@ -47,6 +47,8 @@ def element_to_objects(payload: Dict) -> List:
     entity._source_protocol = "activitypub"
     # Save element object to entity for possible later use
     entity._source_object = payload
+    # Extract receivers
+    entity._receivers = extract_receivers(payload)
 
     if hasattr(entity, "post_receive"):
         entity.post_receive()
@@ -64,6 +66,49 @@ def element_to_objects(payload: Dict) -> List:
     entities.append(entity)
 
     return entities
+
+
+def extract_receiver(payload: Dict, receiver: str) -> Optional[UserType]:
+    """
+    Transform a single receiver ID to a UserType.
+    """
+    if receiver == NAMESPACE_PUBLIC:
+        # Ignore since we already store "public" as a boolean on the entity
+        return
+    # Check for this being a list reference to followers of an actor?
+    # TODO: terrible hack! the way some platforms deliver to sharedInbox using just
+    #   the followers collection as a target is annoying to us since we would have to
+    #   store the followers collection references on application side, which we don't
+    #   want to do since it would make application development another step more complex.
+    #   So for now we're going to do a terrible assumption that
+    #     1) if "followers" in ID and
+    #     2) if ID starts with actor ID
+    #     then; assume this is the followers collection of said actor ID.
+    #   When we have a caching system, just fetch each receiver and check what it is.
+    #   Without caching this would be too expensive to do.
+    elif receiver.find("followers") > -1 and receiver.startswith(payload.get('actor')):
+        return UserType(id=payload.get("actor"), receiver_variant=ReceiverVariant.FOLLOWERS)
+    # Assume actor ID
+    return UserType(id=receiver, receiver_variant=ReceiverVariant.ACTOR)
+
+
+def extract_receivers(payload: Dict) -> List[UserType]:
+    """
+    Exctract receivers from a payload.
+    """
+    receivers = []
+    for key in ("to", "cc"):
+        receiver = payload.get(key)
+        if isinstance(receiver, list):
+            for item in receiver:
+                extracted = extract_receiver(payload, item)
+                if extracted:
+                    receivers.append(extracted)
+        elif isinstance(receiver, str):
+            extracted = extract_receiver(payload, receiver)
+            if extracted:
+                receivers.append(extracted)
+    return receivers
 
 
 def get_outbound_entity(entity: BaseEntity, private_key):